Move selectNextEditable from util to device.
