+++ src/components/Search.tsx
@@ -2,9 +2,10 @@ import React, { useRef } from 'react'
 import { connect } from 'react-redux'
 import _ from 'lodash'
 import { store } from '../store'
-import { selectNextEditable, strip } from '../util'
+import { strip } from '../util'
 import { Connected } from '../@types'
 import { search, setCursor } from '../action-creators'
+import selectNextEditable from '../device/selectNextEditable'
 
 // components
 import ContentEditable, { ContentEditableEvent } from 'react-contenteditable'
