+++ src/device/selectNextEditable.ts
@@ -1,5 +1,5 @@
 /** Focus on the next .editable element in the DOM after the given .editable. May be a sibling or the nearest ancestor's next sibling. */
-export const selectNextEditable = (currentNode: Node) => {
+const selectNextEditable = (currentNode: Node) => {
   const allElements = document.querySelectorAll('.editable')
   const currentIndex = Array.prototype.findIndex.call(allElements, el => currentNode.isEqualNode(el))
   if (currentIndex < allElements.length - 1) {
@@ -7,3 +7,5 @@ export const selectNextEditable = (currentNode: Node) => {
     el.focus()
   }
 }
+
+export default selectNextEditable
