+++ src/util/index.ts
@@ -85,7 +85,6 @@ export { removeContext } from './removeContext'
 export { removeDuplicatedContext } from './removeDuplicatedContext'
 export { removeHome } from './removeHome'
 export { roamJsonToBlocks } from './roamJsonToBlocks'
-export { selectNextEditable } from './selectNextEditable'
 export { setSelection } from './setSelection'
 export { sort } from './sort'
 export { spellNumber } from './spellNumber'
